✨Component refactoring
