+++ src/components/AnnouncementBar.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { defineProps } from 'vue';
 
-const props = defineProps({
+defineProps({
     message: String,
 });
